refactor(ui-browser): extract grid list preview and scrollbar rendering
