+++ src/ui/browser/mod.rs
@@ -1,23 +1,14 @@
 mod entries;
+mod grid;
 mod layout;
+mod list;
+mod preview;
+mod scrollbar;
 mod sidebar;
 
 use super::theme::Palette;
-use super::{helpers, theme};
-use crate::app::{App, Entry, EntryHit, FrameState, ViewMetrics};
-use ratatui::{
-    Frame,
-    buffer::Buffer,
-    layout::{Alignment, Constraint, Direction, Layout, Margin, Rect},
-    style::{Modifier, Style},
-    text::{Line, Span},
-    widgets::{Block, BorderType, Borders, Paragraph, Widget},
-};
-
-use self::entries::{
-    browser_directory_secondary, browser_entry_detail, browser_entry_modified,
-    render_compact_list_row,
-};
+use crate::app::{App, FrameState};
+use ratatui::{Frame, layout::Rect};
 
 pub(super) fn render_body(
     frame: &mut Frame<'_>,
@@ -29,700 +20,14 @@ pub(super) fn render_body(
     layout::render_body(frame, area, app, state, palette);
 }
 
-fn render_grid(
-    frame: &mut Frame<'_>,
-    area: Rect,
-    app: &App,
-    state: &mut FrameState,
-    palette: Palette,
-) {
-    let (content_area, scrollbar_area) = split_scrollbar_area(area);
-
-    helpers::fill_area(frame, content_area, palette.panel_alt, palette.text);
-    if let Some(sb) = scrollbar_area {
-        helpers::fill_area(frame, sb, palette.panel_alt, palette.border);
-    }
-
-    let spec = helpers::grid_zoom_spec(app.zoom_level);
-    let gap_x = spec.gap_x;
-    let gap_y = spec.gap_y;
-    let cols = ((content_area.width + gap_x) / (spec.tile_width_hint + gap_x)).max(1) as usize;
-    let total_gap_x = gap_x.saturating_mul(cols.saturating_sub(1) as u16);
-    let tile_width =
-        (content_area.width.saturating_sub(total_gap_x) / cols as u16).max(spec.min_tile_width);
-    let rows_visible = ((content_area.height + gap_y) / (spec.tile_height + gap_y)).max(1) as usize;
-    state.metrics = ViewMetrics { cols, rows_visible };
-
-    if app.entries.is_empty() {
-        helpers::render_empty_state(frame, content_area, "This folder is empty", palette);
-        return;
-    }
-
-    let start = app.scroll_row * cols;
-    let limit = rows_visible * cols;
-
-    for (visible_index, entry_index) in (start..app.entries.len()).take(limit).enumerate() {
-        let row = visible_index / cols;
-        let col = visible_index % cols;
-        let rect = Rect {
-            x: content_area.x + col as u16 * (tile_width + gap_x),
-            y: content_area.y + row as u16 * (spec.tile_height + gap_y),
-            width: tile_width,
-            height: spec.tile_height,
-        };
-        let entry = &app.entries[entry_index];
-        render_tile(
-            frame,
-            rect,
-            app,
-            entry,
-            entry_index == app.selected,
-            palette,
-            spec,
-        );
-        state.entry_hits.push(EntryHit {
-            rect,
-            index: entry_index,
-        });
-    }
-
-    if let Some(sb) = scrollbar_area {
-        let total_rows = app.entries.len().div_ceil(cols);
-        render_browser_scrollbar(frame, sb, total_rows, rows_visible, app.scroll_row, palette);
-    }
-}
-
-fn render_tile(
-    frame: &mut Frame<'_>,
-    rect: Rect,
-    app: &App,
-    entry: &Entry,
-    selected: bool,
-    palette: Palette,
-    spec: helpers::GridZoomSpec,
-) {
-    let icon_color = theme::entry_color(entry, palette);
-    let background = palette.surface;
-    let content_bg = if selected {
-        theme::mix_color(palette.selected_bg, icon_color, 22)
-    } else {
-        palette.surface
-    };
-    let band_bg = palette.elevated;
-    let band_fg = palette.text;
-    let band_icon = icon_color;
-
-    frame.render_widget(
-        Block::default().style(Style::default().bg(background).fg(palette.text)),
-        rect,
-    );
-
-    let band = Rect {
-        x: rect.x,
-        y: rect.y,
-        width: rect.width,
-        height: 1,
-    };
-    frame.render_widget(
-        Block::default().style(Style::default().bg(band_bg).fg(band_fg)),
-        band,
-    );
-    frame.render_widget(
-        Paragraph::new(Line::from(vec![
-            Span::styled(
-                theme::entry_symbol(entry),
-                Style::default().fg(band_icon).add_modifier(
-                    Modifier::BOLD
-                        | if spec.emphasize_icon {
-                            Modifier::ITALIC
-                        } else {
-                            Modifier::empty()
-                        },
-                ),
-            ),
-            Span::raw(" "),
-            Span::styled(
-                helpers::clamp_label(&entry.name, band.width.saturating_sub(5) as usize),
-                Style::default().fg(band_fg).add_modifier(Modifier::BOLD),
-            ),
-        ]))
-        .style(Style::default().bg(band_bg).fg(band_fg)),
-        band.inner(Margin {
-            horizontal: 1,
-            vertical: 0,
-        }),
-    );
-
-    let content = Rect {
-        x: rect.x,
-        y: rect.y.saturating_add(1),
-        width: rect.width,
-        height: rect.height.saturating_sub(1),
-    };
-    let content_inner = content.inner(Margin {
-        horizontal: spec.padding_x,
-        vertical: 0,
-    });
-    let content_text = Rect {
-        x: content_inner.x,
-        y: content_inner.y,
-        width: content_inner.width,
-        height: content_inner.height,
-    };
-    let detail = browser_entry_detail(app, entry);
-    let modified = browser_entry_modified(entry);
-    let mut lines = Vec::new();
-    if spec.show_kind_hint {
-        lines.push(Line::from(Span::styled(
-            if entry.is_dir() {
-                "Open folder"
-            } else {
-                "Open file"
-            },
-            Style::default().fg(icon_color),
-        )));
-    }
-    if let Some(detail) = detail {
-        lines.push(Line::from(Span::styled(
-            detail,
-            Style::default().fg(palette.muted),
-        )));
-    }
-    lines.push(Line::from(Span::styled(
-        modified,
-        Style::default().fg(palette.muted),
-    )));
-    if content.height > 0 {
-        frame.render_widget(
-            Block::default().style(Style::default().bg(content_bg).fg(palette.text)),
-            content,
-        );
-        frame.render_widget(
-            Paragraph::new(lines).style(Style::default().bg(content_bg).fg(palette.text)),
-            content_text,
-        );
-    }
-}
-
-fn render_list(
-    frame: &mut Frame<'_>,
-    area: Rect,
-    app: &App,
-    state: &mut FrameState,
-    palette: Palette,
-) {
-    let (content_area, scrollbar_area) = split_scrollbar_area(area);
-
-    helpers::fill_area(frame, content_area, palette.panel_alt, palette.text);
-    if let Some(sb) = scrollbar_area {
-        helpers::fill_area(frame, sb, palette.panel_alt, palette.border);
-    }
-
-    let row_height = helpers::list_row_height(app.zoom_level);
-    state.metrics = ViewMetrics {
-        cols: 1,
-        rows_visible: (content_area.height / row_height.max(1)).max(1) as usize,
-    };
-
-    if app.entries.is_empty() {
-        helpers::render_empty_state(frame, content_area, "This folder is empty", palette);
-        return;
-    }
-
-    for (visible_index, entry_index) in (app.scroll_row..app.entries.len())
-        .take(state.metrics.rows_visible)
-        .enumerate()
-    {
-        let entry = &app.entries[entry_index];
-        let row = Rect {
-            x: content_area.x,
-            y: content_area.y + visible_index as u16 * row_height,
-            width: content_area.width,
-            height: row_height,
-        };
-        let selected = entry_index == app.selected;
-        let multi_selected = app.is_selected(&entry.path);
-        let icon_color = theme::entry_color(entry, palette);
-        let bg = if selected {
-            palette.selected_bg
-        } else {
-            palette.panel_alt
-        };
-        if row_height == 1 {
-            frame.render_widget(
-                Paragraph::new(render_compact_list_row(
-                    app, entry, selected, row.width, palette,
-                ))
-                .style(Style::default().bg(bg).fg(palette.text)),
-                row,
-            );
-        } else {
-            let bar_color = if selected {
-                palette.selected_border
-            } else if multi_selected {
-                palette.selection_bar
-            } else {
-                bg
-            };
-            let columns = Layout::default()
-                .direction(Direction::Horizontal)
-                .constraints([Constraint::Length(1), Constraint::Min(1)])
-                .split(row);
-            frame.render_widget(
-                Paragraph::new(if selected || multi_selected {
-                    "▌"
-                } else {
-                    " "
-                })
-                .alignment(Alignment::Left)
-                .style(Style::default().bg(bg).fg(bar_color)),
-                columns[0],
-            );
-            let secondary = if entry.is_dir() {
-                browser_directory_secondary(app, entry)
-            } else if row_height >= 3 {
-                format!(
-                    "{}  •  {}",
-                    browser_entry_detail(app, entry).unwrap_or_default(),
-                    browser_entry_modified(entry)
-                )
-            } else {
-                browser_entry_detail(app, entry).unwrap_or_default()
-            };
-            frame.render_widget(
-                Paragraph::new(vec![
-                    Line::from(vec![
-                        Span::styled(
-                            theme::entry_symbol(entry),
-                            Style::default().fg(icon_color).add_modifier(Modifier::BOLD),
-                        ),
-                        Span::raw(" "),
-                        Span::styled(
-                            helpers::clamp_label(&entry.name, row.width.saturating_sub(8) as usize),
-                            Style::default()
-                                .fg(palette.text)
-                                .add_modifier(Modifier::BOLD),
-                        ),
-                    ]),
-                    Line::from(vec![
-                        Span::raw("  "),
-                        Span::styled(secondary, Style::default().fg(palette.muted)),
-                    ]),
-                ])
-                .style(Style::default().bg(bg).fg(palette.text)),
-                columns[1],
-            );
-        }
-        state.entry_hits.push(EntryHit {
-            rect: row,
-            index: entry_index,
-        });
-    }
-
-    if let Some(sb) = scrollbar_area {
-        render_browser_scrollbar(
-            frame,
-            sb,
-            app.entries.len(),
-            state.metrics.rows_visible,
-            app.scroll_row,
-            palette,
-        );
-    }
-}
-
-fn render_preview(
-    frame: &mut Frame<'_>,
-    area: Rect,
-    app: &App,
-    state: &mut FrameState,
-    palette: Palette,
-) {
-    state.preview_panel = Some(area);
-
-    let title_line = if let Some(entry) = app.selected_entry() {
-        Line::from(vec![
-            Span::styled(
-                format!(" {} ", theme::entry_symbol(entry)),
-                Style::default()
-                    .fg(theme::entry_color(entry, palette))
-                    .add_modifier(Modifier::BOLD),
-            ),
-            Span::styled(
-                helpers::clamp_label(&entry.name, area.width.saturating_sub(10) as usize),
-                Style::default()
-                    .fg(palette.accent_text)
-                    .add_modifier(Modifier::BOLD),
-            ),
-            Span::raw(" "),
-        ])
-    } else {
-        Line::from(vec![
-            Span::styled(
-                " Preview ",
-                Style::default()
-                    .fg(palette.accent_text)
-                    .add_modifier(Modifier::BOLD),
-            ),
-            Span::raw(" "),
-        ])
-    };
-    let block = Block::default()
-        .borders(Borders::ALL)
-        .border_type(BorderType::Rounded)
-        .style(Style::default().bg(palette.panel).fg(palette.text))
-        .border_style(Style::default().fg(palette.border));
-    frame.render_widget(block, area);
-    helpers::render_panel_title(frame, area, title_line);
-    let inner = helpers::inner_with_padding(area);
-    helpers::fill_area(frame, inner, palette.panel, palette.text);
-
-    let Some(entry) = app.selected_entry() else {
-        helpers::render_empty_state(frame, inner, "Nothing selected", palette);
-        return;
-    };
-
-    if inner.height > 0 {
-        render_preview_body(frame, inner, app, state, entry, palette);
-    }
-}
-
-fn render_preview_body(
-    frame: &mut Frame<'_>,
-    area: Rect,
-    app: &App,
-    state: &mut FrameState,
-    _entry: &Entry,
-    palette: Palette,
-) {
-    let sections = Layout::default()
-        .direction(Direction::Vertical)
-        .constraints([Constraint::Length(1), Constraint::Min(0)])
-        .split(area);
-    state.preview_rows_visible = sections[1].height as usize;
-    helpers::fill_area(frame, sections[0], palette.panel, palette.text);
-    if sections[1].height > 0 {
-        helpers::fill_area(frame, sections[1], palette.panel, palette.text);
-    }
-    let body = if sections[1].width >= 6 {
-        Layout::default()
-            .direction(Direction::Horizontal)
-            .constraints([Constraint::Min(0), Constraint::Length(1)])
-            .split(sections[1])
-    } else {
-        Layout::default()
-            .direction(Direction::Horizontal)
-            .constraints([Constraint::Min(0)])
-            .split(sections[1])
-    };
-    let body_area = body[0];
-    let scrollbar_area = body.get(1).copied();
-    state.preview_body_area = Some(sections[1]);
-    let (media_area, text_area) = if let Some(media_rows) = app.preview_visual_rows(body_area) {
-        let split = Layout::default()
-            .direction(Direction::Vertical)
-            .constraints([Constraint::Length(media_rows), Constraint::Min(0)])
-            .split(body_area);
-        (Some(split[0]), split[1])
-    } else {
-        (None, body_area)
-    };
-    state.preview_media_area = media_area;
-    state.preview_content_area = Some(text_area);
-    if let Some(media_area) = media_area {
-        helpers::fill_area(frame, media_area, palette.panel, palette.text);
-    }
-    helpers::fill_area(frame, text_area, palette.panel, palette.text);
-    if let Some(scrollbar_area) = scrollbar_area {
-        helpers::fill_area(frame, scrollbar_area, palette.panel, palette.border);
-    }
-    let visible_rows = text_area.height as usize;
-    state.preview_cols_visible = text_area.width as usize;
-    let section_label = app.preview_section_label();
-    let header_detail_width = sections[0]
-        .width
-        .saturating_sub(section_label.len() as u16 + 2) as usize;
-    let header_detail = app
-        .preview_header_detail_for_width(visible_rows, header_detail_width)
-        .as_deref()
-        .map(|detail| {
-            if header_detail_width == 0 {
-                String::new()
-            } else {
-                helpers::clamp_label(detail, header_detail_width)
-            }
-        })
-        .unwrap_or_default();
-
-    frame.render_widget(
-        Paragraph::new(Line::from(vec![
-            Span::styled(
-                section_label.to_string(),
-                Style::default()
-                    .fg(palette.accent)
-                    .add_modifier(Modifier::BOLD),
-            ),
-            Span::styled("  ", Style::default().fg(palette.muted)),
-            Span::styled(header_detail, Style::default().fg(palette.muted)),
-        ]))
-        .style(Style::default().bg(palette.panel).fg(palette.text)),
-        sections[0],
-    );
-
-    if app.browser_wheel_burst_active() {
-        frame.render_widget(
-            Paragraph::new(Line::from(Span::styled(
-                "Scrolling...",
-                Style::default().fg(palette.muted),
-            )))
-            .style(Style::default().bg(palette.panel).fg(palette.text))
-            .alignment(Alignment::Center),
-            text_area,
-        );
-        return;
-    }
-
-    if app.preview_prefers_image_surface() {
-        if let Some(message) = app.preview_overlay_placeholder_message() {
-            frame.render_widget(
-                Paragraph::new(Line::from(Span::styled(
-                    message,
-                    Style::default().fg(palette.muted),
-                )))
-                .style(Style::default().bg(palette.panel).fg(palette.text))
-                .alignment(Alignment::Center),
-                text_area,
-            );
-        }
-        return;
-    }
-
-    if app.preview_uses_image_overlay() {
-        return;
-    }
-
-    if let Some(media_area) = media_area
-        && let Some(message) = app.preview_visual_placeholder_message()
-    {
-        frame.render_widget(
-            Paragraph::new(Line::from(Span::styled(
-                message,
-                Style::default().fg(palette.muted),
-            )))
-            .style(Style::default().bg(palette.panel).fg(palette.text))
-            .alignment(Alignment::Center),
-            media_area,
-        );
-    }
-
-    if app.preview_wraps() {
-        let wrapped_lines = app.preview_wrapped_lines(text_area.width as usize);
-        frame.render_widget(
-            PreviewLinesWidget::new(
-                wrapped_lines.as_ref(),
-                app.preview_scroll_offset(),
-                Style::default().bg(palette.panel).fg(palette.text),
-            ),
-            text_area,
-        );
-    } else {
-        let paragraph = Paragraph::new(app.preview_lines())
-            .style(Style::default().bg(palette.panel).fg(palette.text))
-            .scroll((
-                app.preview_scroll_offset().min(u16::MAX as usize) as u16,
-                app.preview_horizontal_scroll_offset()
-                    .min(u16::MAX as usize) as u16,
-            ));
-        frame.render_widget(paragraph, text_area);
-    }
-
-    if let Some(scrollbar_area) = scrollbar_area {
-        render_preview_scrollbar(
-            frame,
-            scrollbar_area,
-            app,
-            visible_rows,
-            text_area.width as usize,
-            palette,
-        );
-    }
-}
-
-struct PreviewLinesWidget<'a> {
-    lines: &'a [Line<'static>],
-    scroll: usize,
-    style: Style,
-}
-
-impl<'a> PreviewLinesWidget<'a> {
-    fn new(lines: &'a [Line<'static>], scroll: usize, style: Style) -> Self {
-        Self {
-            lines,
-            scroll,
-            style,
-        }
-    }
-}
-
-impl Widget for PreviewLinesWidget<'_> {
-    fn render(self, area: Rect, buf: &mut Buffer) {
-        let area = area.intersection(buf.area);
-        if area.is_empty() {
-            return;
-        }
-
-        buf.set_style(area, self.style);
-        for (line, row) in self.lines.iter().skip(self.scroll).zip(area.rows()) {
-            let line_width = line.width();
-            let offset = match line.alignment.unwrap_or(Alignment::Left) {
-                Alignment::Center => row.width.saturating_sub(line_width as u16) / 2,
-                Alignment::Right => row.width.saturating_sub(line_width as u16),
-                Alignment::Left => 0,
-            };
-            if offset >= row.width {
-                continue;
-            }
-
-            let x = row.x.saturating_add(offset);
-            let max_width = row.width.saturating_sub(offset);
-            buf.set_line(x, row.y, line, max_width);
-        }
-    }
-}
-
-fn render_preview_scrollbar(
-    frame: &mut Frame<'_>,
-    area: Rect,
-    app: &App,
-    visible_rows: usize,
-    visible_cols: usize,
-    palette: Palette,
-) {
-    let total = app.preview_total_lines(visible_cols);
-    if area.height == 0 || total <= visible_rows.max(1) {
-        frame.render_widget(
-            Paragraph::new(" ").style(Style::default().bg(palette.panel).fg(palette.border)),
-            area,
-        );
-        return;
-    }
-
-    let track = vec![
-        Line::from(Span::styled("│", Style::default().fg(palette.border),));
-        area.height as usize
-    ];
-    frame.render_widget(
-        Paragraph::new(track).style(Style::default().bg(palette.panel)),
-        area,
-    );
-
-    let thumb_height = ((visible_rows.max(1) * area.height as usize) / total)
-        .max(1)
-        .min(area.height as usize);
-    let max_scroll = total.saturating_sub(visible_rows.max(1));
-    let thumb_max_top = area.height as usize - thumb_height;
-    let thumb_top = if max_scroll == 0 {
-        0
-    } else {
-        (app.preview_scroll_offset() * thumb_max_top) / max_scroll
-    };
-
-    let thumb = Rect {
-        x: area.x,
-        y: area.y + thumb_top as u16,
-        width: area.width,
-        height: thumb_height as u16,
-    };
-    frame.render_widget(
-        Paragraph::new(vec![
-            Line::from(Span::styled(
-                "┃",
-                Style::default()
-                    .fg(palette.accent)
-                    .add_modifier(Modifier::BOLD),
-            ));
-            thumb.height as usize
-        ])
-        .style(Style::default().bg(palette.panel)),
-        thumb,
-    );
-}
-
-fn split_scrollbar_area(area: Rect) -> (Rect, Option<Rect>) {
-    if area.width >= 6 {
-        let parts = Layout::default()
-            .direction(Direction::Horizontal)
-            .constraints([Constraint::Min(0), Constraint::Length(1)])
-            .split(area);
-        (parts[0], Some(parts[1]))
-    } else {
-        (area, None)
-    }
-}
-
-fn render_browser_scrollbar(
-    frame: &mut Frame<'_>,
-    area: Rect,
-    total_rows: usize,
-    visible_rows: usize,
-    scroll_row: usize,
-    palette: Palette,
-) {
-    if area.height == 0 || total_rows <= visible_rows.max(1) {
-        frame.render_widget(
-            Paragraph::new(" ").style(Style::default().bg(palette.panel_alt).fg(palette.border)),
-            area,
-        );
-        return;
-    }
-
-    let track = vec![
-        Line::from(Span::styled("│", Style::default().fg(palette.border)));
-        area.height as usize
-    ];
-    frame.render_widget(
-        Paragraph::new(track).style(Style::default().bg(palette.panel_alt)),
-        area,
-    );
-
-    let thumb_height = ((visible_rows.max(1) * area.height as usize) / total_rows)
-        .max(1)
-        .min(area.height as usize);
-    let max_scroll = total_rows.saturating_sub(visible_rows.max(1));
-    let thumb_max_top = area.height as usize - thumb_height;
-    let thumb_top = if max_scroll == 0 {
-        0
-    } else {
-        (scroll_row * thumb_max_top) / max_scroll
-    };
-
-    let thumb = Rect {
-        x: area.x,
-        y: area.y + thumb_top as u16,
-        width: area.width,
-        height: thumb_height as u16,
-    };
-    frame.render_widget(
-        Paragraph::new(vec![
-            Line::from(Span::styled(
-                "┃",
-                Style::default()
-                    .fg(palette.accent)
-                    .add_modifier(Modifier::BOLD),
-            ));
-            thumb.height as usize
-        ])
-        .style(Style::default().bg(palette.panel_alt)),
-        thumb,
-    );
-}
-
 #[cfg(test)]
 mod tests {
-    use super::*;
+    use super::super::theme;
+    use super::scrollbar::split_scrollbar_area;
+    use crate::app::{App, FrameState};
     use crate::ui;
     use crossterm::event::{Event, KeyCode, KeyEvent, KeyModifiers};
-    use ratatui::{Terminal, backend::TestBackend, buffer::Buffer};
+    use ratatui::{Terminal, backend::TestBackend, buffer::Buffer, layout::Rect, style::Modifier};
     use std::{
         fs,
         path::PathBuf,
